requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pandas
Pillow
pdfplumber
pdf2image
#yolov5
opencv-python

PyYAML
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
seaborn
python-multipart
psutil  # system resources

requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation

tqdm>=4.64.0
